% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadGISFiles.R
\name{LOAD.AND.PREP.GIS}
\alias{LOAD.AND.PREP.GIS}
\title{Load and Prep a Shapefile}
\usage{
LOAD.AND.PREP.GIS(SHPNAME, PROJECT.PATH = PROJECT.PATH,
  PROJ.USE = PROJ.USE)
}
\arguments{
\item{SHPNAME}{the name of the shapefile to be loaded}

\item{PROJECT.PATH}{the filepath of the shapefile}

\item{PROJ.USE}{the \code{\link{CRS}} projection of the shapefile}
}
\value{
The output is a Large SpatialPolygon S4 object
}
\description{
This function loads a shapefile and preps it so that it can be used in the function
}
\examples{
LOAD.AND.PREP.GIS("clipArea", "C:/Users/dennis.corvi/Documents/R/Projects", albersProj)

}
